+++ src/plugin.js
@@ -5,27 +5,36 @@ var old = $.fn.bootgrid;
 
 $.fn.bootgrid = function (option)
 {
-    var args = Array.prototype.slice.call(arguments, 1);
-    return this.each(function ()
-    {
-        var $this = $(this),
-            instance = $this.data(namespace),
-            options = typeof option === "object" && option;
+    var args = Array.prototype.slice.call(arguments, 1),
+        returnValue = null,
+        elements = this.each(function (index)
+        {
+            var $this = $(this),
+                instance = $this.data(namespace),
+                options = typeof option === "object" && option;
 
-        if (!instance && option === "destroy")
-        {
-            return;
-        }
-        if (!instance)
-        {
-            $this.data(namespace, (instance = new Grid(this, options)));
-            init.call(instance);
-        }
-        if (typeof option === "string")
-        {
-            return instance[option].apply(instance, args);
-        }
-    });
+            if (!instance && option === "destroy")
+            {
+                return;
+            }
+            if (!instance)
+            {
+                $this.data(namespace, (instance = new Grid(this, options)));
+                init.call(instance);
+            }
+            if (typeof option === "string")
+            {
+                if (option.indexOf("get") === 0 && index === 0)
+                {
+                    returnValue = instance[option].apply(instance, args);
+                }
+                else if (option.indexOf("get") !== 0)
+                {
+                    return instance[option].apply(instance, args);
+                }
+            }
+        });
+    return (typeof option === "string" && option.indexOf("get") === 0) ? returnValue : elements;
 };
 
 $.fn.bootgrid.Constructor = Grid;